Create abstract class AffectsShroud[Info] with RevealsShroud core logic also used by CreatesShroud
